Model all `int` as `Long` in JetBrains generation
Service allows `int` to be 64-bit, so generate these as `Long` for consistency across products

+++ telemetry/jetbrains/src/main/kotlin/software/aws/toolkits/telemetry/generator/TelemetryParser.kt
@@ -18,7 +18,7 @@ enum class MetricMetadataTypes(@get:JsonValue val type: String) {
         override fun kotlinType(): TypeName = com.squareup.kotlinpoet.STRING
     },
     INT("int") {
-        override fun kotlinType(): TypeName = com.squareup.kotlinpoet.INT
+        override fun kotlinType(): TypeName = com.squareup.kotlinpoet.LONG
     },
     DOUBLE("double") {
         override fun kotlinType(): TypeName = com.squareup.kotlinpoet.DOUBLE
